Bluetooth: Avoiding a ANR, if FTP service record is not registered
A case where monkey is running. When BT is turned on, Rfcomm
Server Socket is not created during FTP SDP registration on 1st
attempt. So FTP was retrying to create RFCOMM server socket after
sleeping 3 second. As it was running under main thread context. So
ANR was occurred due to longer sleep(3s). Reducing the sleep from 3s
to 300ms to solve the issue.

Change-Id: I93b9e6f77dc1a1f3d941571299e86285f0b44b32
CRs-Fixed: 611677