Bluetooth: Pxpmonitor DeviceActivity Crash When HOGP Device Connected
DeviceActivity in Pxpmonitor is destoryed and restarted by Android
framework on change of configuration. During restart there is delay
in getting the Service Connection status which return the Proxy
Service Handle on successful connection. This Proxy Service handle
was deferenced without null check. Added check to deference
proxyService handle only when it is available.

Change-Id: Ic67549e34289181bafbdae5d3a98113177aa9981
CRs-Fixed: 738143