PBAPC: Do abort when onDestroy() is called
This change will abort the current ongoing phonebook
download when going back to the Main Activity from the
Phonebook activity since the Pbap Activity is no longer active.
Earlier this was happening on onPause() but it was creating issues
since onPause() is called in other scenarios also eg. when device goes
into sleep.

CRs-Fixed: 787844
Change-Id: I7c37ad708d203664a70b7fd0586ce1899cc4b1d0