Fix DUT1 as sink fail to connect with DUT2
DUT1 as sink to connect DUT2, DUT2 will check the COD value find
it is not a sink_only or src_only device, and will initial the src
and sink connet connect. When src is configured and it will dis-
connect the unused sink, but will call function dealloc_ar_device_
info to remove the device info, which will cause problem in the
next connect operation. The fix will check the sep info, if there
is src connection exit, will not dealloct the device.

Change-Id: I18671b1191ea65b2d5b3070d3860d2018676ce1e
CRs-Fixed: 2075857