Write descriptor is getting failed sometimes in packages.
When two clients are doing bulk tx to the same remote using 'TxRx' option in throughput menu,
after the completion of transfer we deregister notifications,
write descriptor is getting failed in packages layer as mDeviceBusyLock is not being released.
The same lock is used for write characteristic as well, since the second app
is writing to characteristic continuously, first app may not get a chance to write
to descriptor to disable notifications at the end of the test.

There is no retry mechanism for write descriptor in packages layer, hence
adding retry mechanism for write descriptor after a delay.

Change-Id: Ie8b2374a8d34c53a579de066daa7063059427a5c